Bump AWS SDK version to `2.13.8`
Rewrites this interceptor to use AWS SDK version 2.x.x. Renames
interceptor to use `Aws` instead of `AWS` in the name as classes do on
SDK 2. Bumps JUnit to version 5. Change group to be able to deploy this
library to a Maven repo given AWS is not doing so at the moment.