Fix: signing POST requests (#31)
* Fix: POST requests to AWS OpenSearch.
* Added gzip data sample and tests.
* Added unit test for HTTP POST signing.
* Use IoUtils to read stream length.
* Cleanup stream creation, raise error on HTTP failures.
* Use entity classes directly.

The interceptor consumes a stream of data, then copies the data back to the original request. The fix is in the interceptor code: create a new stream for the data to avoid having consumed it all (there would be nothing left to copy when copying back). Added a unit test for HTTP POST that covers checking that the content is still present after signing the data.

The second change is to correctly write the JSON string and properly calculate its content length. Otherwise we end up sending \0 along for the ride. That's only a bug in the HTTP demo, thus the @acm19 version described in #29 was working. I also added demos for gzipped content and failing demos for the same with chunked transfer encoding.

Signed-off-by: Daniel (dB.) Doubrovkine <dblock@amazon.com>