Automate release process
Fully automates the release process by moving from a Makefile that had
to be executed locally to a job that is triggered in GitHub. This
removes possible errors and manual burden associated to releasing.

The new process skips the creation of intermediate branches and pull
requests and commits directly to master, which then triggers the
actual pushing of the artefacts together with the tagging of the repo
the same way it was done before this change.

Resolves: https://github.com/acm19/aws-request-signing-apache-interceptor/issues/53