Add support for Apache client v5 (#80)
Adds support for Apache client v5. Users can simply add the Apache
client dependency they need and use the class according to that version
to get their requests signed.

Adjusts the checkstyle so it only run in a profile. Changes the style of
the import of dependencies to be consistent with the one used by the AWS
SDK.

Resolves: https://github.com/acm19/aws-request-signing-apache-interceptor/issues/77