Make request interceptors simpler and other improvements
Reduces code in the interceptor by setting back the body at the same
time it's retrieve. Since the body is not modified during the signing
process it's not necessary to have the same conditions twice in the
code.

Removes check for `final` parameters as it mostly makes the code more
verbose and it helps very little in such a small library.

Removes use of deprecated `httpClient.execute` methods replacing it by
the currently supported versions of the same method, updates the
documentation accordingly.

Bumps the version of the testing OpenSearch cluster to 2.5.