Automate the release fully by auto closing it
Adds the `nexus-staging-maven-plugin` to fully automate the release
process by autoclosing the release from the pipeline directly.

Enables contributors to merge and release code without manual action in
Maven Central.

Moves Javadoc and release related tasks to a profile to reduce build
time when not releasing.

Removes deprecated command in GitHub job `set-output`, support for it
will be removed soon and the jobs using it will start to fail.

Resolves: https://github.com/acm19/aws-request-signing-apache-interceptor/issues/84.